add library/delete endpoint and complete mangaDetailsDto. Refactors code for library and config files
