+++ src/library/library.service.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { LibraryService } from './library.service';
+
+describe('LibraryService', () => {
+  let service: LibraryService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [LibraryService],
+    }).compile();
+
+    service = module.get<LibraryService>(LibraryService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
