+++ src/mangas/constants.ts
@@ -5,3 +5,17 @@ export const MAL_DETAIL_URL = 'https://api.myanimelist.net/v2/manga/';
 export const MU_TRENDS_URL = 'https://api.mangaupdates.com/v1/series/search';
 
 export const MU_DETAIL_URL = 'https://api.mangaupdates.com/v1/series/';
+
+export const NSFW_GENRES = [
+  'Mature',
+  'Adult',
+  'Smut',
+  'Hentai',
+  'Lolicon',
+  'Shotacon',
+  'Yaoi',
+  'Ecchi',
+  'Yuri',
+  'Harem',
+  'Doujinshi',
+];
