+++ src/mangas/dto/manga-details.dto.ts
@@ -1,38 +1,61 @@
 import { plainToClass, classToPlain } from 'class-transformer';
 import { Manga } from '../manga.entity';
 import { IsBoolean, IsNumber, IsOptional, IsString } from 'class-validator';
+import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
 
 export class MangaDetailsDto {
+  @ApiProperty()
   @IsNumber()
   muId: number;
 
+  @ApiProperty()
   @IsString()
   title: string;
 
+  @ApiProperty()
   @IsString()
   description: string;
 
+  @ApiProperty()
   @IsNumber()
   year: number;
 
+  @ApiProperty()
   @IsString()
   smallCoverUrl: string;
 
+  @ApiProperty()
   @IsString()
   mediumCoverUrl: string;
 
+  @ApiProperty()
   @IsNumber()
   rating: number;
 
+  @ApiPropertyOptional()
   @IsNumber()
   @IsOptional()
   readChapters: number;
 
+  @ApiProperty()
   @IsNumber()
   totalChapters: number;
 
+  @ApiProperty()
   @IsBoolean()
   completed: boolean;
+
+  @ApiPropertyOptional()
+  authors: any[];
+
+  @ApiPropertyOptional()
+  genres: any[];
+
+  @ApiPropertyOptional()
+  anime: any[];
+
+  @ApiPropertyOptional()
+  categories: any[];
 
   static toModel(mangaDetailsDto: MangaDetailsDto): Manga {
     const data = classToPlain(mangaDetailsDto);
@@ -50,6 +73,10 @@ export class MangaDetailsDto {
     mangaDetailsDto['totalChapters'] = muObject['latest_chapter'];
     mangaDetailsDto['completed'] = muObject['completed'];
     mangaDetailsDto['muId'] = muObject['series_id'];
+    mangaDetailsDto['authors'] = muObject['authors'];
+    mangaDetailsDto['genres'] = muObject['genres'];
+    mangaDetailsDto['anime'] = muObject['anime'];
+    mangaDetailsDto['categories'] = muObject['categories'];
     return mangaDetailsDto;
   }
 }
