+++ src/mangas/dto/manga-quick-view.dto.ts
